requirements.txt
hydra-core==2.5
marshmallow_dataclass==8.4.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
pytest==6.2.3
PyYAML==5.4.1
scikit_learn==0.24.2
